[refactor] #31 shareWithImage 함수명 변경 및 확장 구현
- 함수의 재활용성을 고려해 Extension 으로 구현
